#42 Add clearing background
- rename shoaldrag to shoal
- redefine canvas hierarchy
- add clearWater function
