+++ podwodny_swiat_tabliczki_mnozenia/Assets/Scritps/Multiplication.cs
@@ -36,7 +36,8 @@ public class Multiplication : MonoBehaviour
     {
         if (answer.text == (multiplicand * multiplier).ToString())
         {
-            error.text = "Bardzo dobrze "+Globals.playerName+" ! Następne równanie.";
+            error.text = "Bardzo dobrze! Następne równanie.";
+            GameController.clearWater();
             RandomEquation();
         }
         else
