+++ podwodny_swiat_tabliczki_mnozenia/Assets/Scritps/Shoal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 005ae855e3fbd1e42b1f15d0a62f0110
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
