+++ podwodny_swiat_tabliczki_mnozenia/Library/StateCache/SceneView/2b2796-mainStage.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":true,"pivot":{"x":271.5367126464844,"y":-173.4812469482422,"z":-137.46270751953126},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":466.8428649902344,"orthographic":true}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":true,"pivot":{"x":350.7655029296875,"y":55.17090606689453,"z":-164.28712463378907},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":311.4768371582031,"orthographic":true}
