+++ podwodny_swiat_tabliczki_mnozenia/Library/shadercompiler-UnityShaderCompiler.exe0.log
@@ -1,2 +1,2 @@
-Base path: 'C:/Users/W/Desktop/Unity/Editor/Data', plugins path 'C:/Users/W/Desktop/Unity/Editor/Data/PlaybackEngines'
+Base path: 'C:/Program Files/Unity/Hub/Editor/2019.3.9f1/Editor/Data', plugins path 'C:/Program Files/Unity/Hub/Editor/2019.3.9f1/Editor/Data/PlaybackEngines'
 Cmd: initializeCompiler
